wire: move init-message to InitMessage
