+++ packages/lntools-wire/lib/messages/InitMessage.ts
